Api 2

+++ product/product.routes.ts
@@ -1,6 +1,7 @@
 import express, {Router} from 'express'
 
 import {Product} from './product.schema'
+import passport from "../middleware/passport";
 
 export const router = Router()
 
@@ -26,51 +27,63 @@ router.get('/:id', async (req, res) => {
     }
 })
 
-router.post('/',
+router.post('/', passport.authenticate('jwt', { session: false }),
     async (req: express.Request, res: express.Response) => {
         try {
-            const existingProduct = await Product.findOne({name: req.body.name})
-            if (existingProduct) {
-                return res.status(400).json({message: 'Such product exists'})
+            if (req.user?.role === 'admin') {
+                const existingProduct = await Product.findOne({name: req.body.name})
+                if (existingProduct) {
+                    return res.status(400).json({message: 'Such product exists'})
+                }
+
+                const product = new Product(req.body)
+                await product.save()
+
+                res.status(201).json('Product added')
+            } else {
+                res.status(403).json({message: 'You don\'t have permission to access this resource'})
             }
-
-            const product = new Product(req.body)
-            await product.save()
-
-            res.status(201).json('Product added')
-
         } catch (e) {
             res.status(400).json({message: 'Incorrect data'})
         }
     })
 
-router.patch('/:id',
+router.patch('/:id', passport.authenticate('jwt', { session: false }),
     async (req: express.Request, res: express.Response) => {
         try {
-            const params = req.body
-            const product = await Product.findById(req.params.id)
+            if (req.user?.role === 'admin') {
+                const params = req.body
+                const product = await Product.findById(req.params.id)
 
-            if (!product) {
-                res.status(404).json({message: 'Product does not exist'})
-                return
+                if (!product) {
+                    res.status(404).json({message: 'Product does not exist'})
+                    return
+                }
+                Object.assign(product, params)
+                await product.save()
+                res.status(200).json('Product updated')
+            } else {
+                res.status(403).json({message: 'You don\'t have permission to access this resource'})
             }
-            Object.assign(product, params)
-            await product.save()
-            res.status(200).json('Product updated')
         } catch (e) {
             res.status(404).json({message: 'Bad request'})
         }
     })
 
-router.delete('/:id', async (req, res) => {
+router.delete('/:id', passport.authenticate('jwt', { session: false }),
+    async (req, res) => {
     try {
-        const product = await Product.findById(req.params.id)
-        if (!product) {
-            res.status(404).json({message: 'Product does not exist'})
-            return
+        if (req.user?.role === 'admin') {
+            const product = await Product.findById(req.params.id)
+            if (!product) {
+                res.status(404).json({message: 'Product does not exist'})
+                return
+            }
+            await product.deleteOne()
+            res.status(200).json('Product deleted')
+        } else {
+            res.status(403).json({message: 'You don\'t have permission to access this resource'})
         }
-        await product.deleteOne()
-        res.status(200).json('Product deleted')
     } catch (e) {
         res.status(404).json({message: 'Bad request'})
     }
