+++ user/user.routes.ts
@@ -1,27 +1,38 @@
 import express, {Router} from 'express'
-import bcrypt from 'bcryptjs'
 
 import {User} from './user.schema'
+import jwt from "jsonwebtoken"
+import passport from "../middleware/passport"
 
 export const router = Router()
 
-router.get('/', async (req, res) => {
+router.get('/', passport.authenticate('jwt', { session: false }),
+    async (req, res) => {
     try {
-        const users = await User.find()
-        res.json(users)
+        if (req.user?.role === 'admin') {
+            const users = await User.find()
+            res.json(users)
+        } else {
+            res.status(403).json({message: 'You don\'t have permission to access this resource'})
+        }
     } catch (e) {
         res.status(500).json({message: 'something went wrong'})
     }
 })
 
-router.get('/:id', async (req, res) => {
+router.get('/:id', passport.authenticate('jwt', { session: false }),
+    async (req, res) => {
     try {
-        const user = await User.findById(req.params.id)
-        if (!user) {
-            res.status(404).json({message: 'User does not exist'})
-            return
+        if (req.user?.role === 'admin' || JSON.stringify(req.params.id) !== JSON.stringify(req.user?._id)) {
+            const user = await User.findById(req.params.id)
+            if (!user) {
+                res.status(404).json({message: 'User does not exist'})
+                return
+            }
+            res.status(200).json(user)
+        } else {
+            res.status(403).json({message: 'You don\'t have permission to access this resource'})
         }
-        res.status(200).json(user)
     } catch (e) {
         res.status(404).json(e)
     }
@@ -31,51 +42,64 @@ router.post('/',
     async (req: express.Request, res: express.Response) => {
         try {
             const {name, phone, email, password} = req.body
-
             const existingUser = await User.findOne({phone})
             const existingUser2 = await User.findOne({email})
-            if (existingUser || existingUser2) {
+            const existingUser3 = await User.findOne({name})
+            if (existingUser || existingUser2 || existingUser3) {
                 return res.status(400).json({message: 'Such user exists'})
             }
 
-            const hashedPassword = await bcrypt.hash(password, 12)
-
-            const user = new User({name, phone, email, password: hashedPassword})
+            const user = new User({name, phone, email, password, role: 'user'})
             await user.save()
 
-            res.status(201).json('User registered')
+            const JWTKey = process.env.JWT_SECRET
+            const body = { _id: user._id, email: user.email }
+            const token = jwt.sign({ user: body }, JWTKey!)
+
+            return res.status(200).json({ token })
         } catch (e) {
+            console.log(e)
             res.status(400).json({message: 'Incorrect data'})
         }
     })
 
-router.patch('/:id',
+router.patch('/:id', passport.authenticate('jwt', { session: false }),
     async (req: express.Request, res: express.Response) => {
         try {
-            const params = req.body
-            const user = await User.findById(req.params.id)
+            if (req.user?.role === 'admin' || JSON.stringify(req.params.id) !== JSON.stringify(req.user?._id)) {
+                const params = req.body
+                const user = await User.findById(req.params.id)
 
-            if (!user) {
-                res.status(404).json({message: 'User does not exist'})
-                return
+                if (!user) {
+                    res.status(404).json({message: 'User does not exist'})
+                    return
+                }
+
+                Object.assign(user, params)
+                await user.save()
+                res.status(200).json('User updated')
+            } else {
+                res.status(403).json({message: 'You don\'t have permission to access this resource'})
             }
-            Object.assign(user, params)
-            await user.save()
-            res.status(200).json('User updated')
         } catch (e) {
             res.status(404).json({message: 'Bad request'})
         }
     })
 
-router.delete('/:id', async (req, res) => {
+router.delete('/:id', passport.authenticate('jwt', { session: false }),
+    async (req, res) => {
     try {
-        const user = await User.findById(req.params.id)
-        if (!user) {
-            res.status(404).json({message: 'User does not exist'})
-            return
+        if (req.user?.role === 'admin') {
+            const user = await User.findById(req.params.id)
+            if (!user) {
+                res.status(404).json({message: 'User does not exist'})
+                return
+            }
+            await user.deleteOne()
+            res.status(200).json('User deleted')
+        } else {
+            res.status(403).json({message: 'You don\'t have permission to access this resource'})
         }
-        await user.deleteOne()
-        res.status(200).json('User deleted')
     } catch (e) {
         res.status(404).json({message: 'Bad request'})
     }
